+++ end_to_end_tests/golden-master/my_test_api_client/api/tests.py
@@ -14,9 +14,7 @@ from ..models.http_validation_error import HTTPValidationError
 
 def get_user_list(
     *, client: Client, an_enum_value: List[AnEnum], some_date: Union[datetime.date, datetime.datetime],
-) -> Union[
-    List[AModel], HTTPValidationError,
-]:
+) -> Union[List[AModel], HTTPValidationError]:
 
     """ Get a list of things  """
     url = "{}/tests/".format(client.base_url)
@@ -112,9 +110,7 @@ def get_basic_list_of_booleans(*, client: Client,) -> List[bool]:
 
 def upload_file_tests_upload_post(
     *, client: Client, multipart_data: BodyUploadFileTestsUploadPost, keep_alive: Optional[bool] = None,
-) -> Union[
-    None, HTTPValidationError,
-]:
+) -> Union[None, HTTPValidationError]:
 
     """ Upload a file  """
     url = "{}/tests/upload".format(client.base_url)
@@ -133,11 +129,7 @@ def upload_file_tests_upload_post(
         raise ApiResponseError(response=response)
 
 
-def json_body_tests_json_body_post(
-    *, client: Client, json_body: AModel,
-) -> Union[
-    None, HTTPValidationError,
-]:
+def json_body_tests_json_body_post(*, client: Client, json_body: AModel,) -> Union[None, HTTPValidationError]:
 
     """ Try sending a JSON body  """
     url = "{}/tests/json_body".format(client.base_url)
@@ -171,9 +163,7 @@ def test_defaults_tests_test_defaults_post(
     ),
     union_prop: Optional[Union[Optional[float], Optional[str]]] = "not a float",
     enum_prop: Optional[AnEnum] = None,
-) -> Union[
-    None, HTTPValidationError,
-]:
+) -> Union[None, HTTPValidationError]:
 
     """  """
     url = "{}/tests/test_defaults".format(client.base_url)