+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.5.5 - Unreleased
+### Fixes
+- Improved trailing comma handling in endpoint generation (#178 & #179). Thanks @dtkav!
+
+
 ## 0.5.4 - 2020-08-29
 ### Additions
 - Support for octet-stream content type (#116)